+++ lib/src/features/purchases/application/use_cases/start_purchase.dart
@@ -1,0 +1,11 @@
+import 'package:repforge/src/features/purchases/domain/purchases_domain.dart';
+
+final class StartPurchase {
+  const StartPurchase(this._gateway);
+
+  final PurchaseGateway _gateway;
+
+  Future<void> call(PurchaseProduct product) {
+    return _gateway.buy(product);
+  }
+}