+++ lib/src/features/purchases/data/purchases_data.dart
@@ -1,0 +1,1 @@
+export 'gateways/in_app_purchase_gateway.dart';