+++ lib/src/features/purchases/domain/purchases_domain.dart
@@ -1,0 +1,7 @@
+export 'entities/purchase_event.dart';
+export 'entities/purchase_product.dart';
+export 'exceptions/purchase_gateway_exception.dart';
+export 'exceptions/purchase_validation_exception.dart';
+export 'repositories/purchase_gateway.dart';
+export 'value_objects/purchase_error.dart';
+export 'value_objects/purchase_product_id.dart';